Removing the GW3 Context factory
Deprecating the gw3 file format. The json format will be the main file
container for graphwalker graphs in the future.
