+++ graphwalker-java/src/test/resources/org/graphwalker/java/test/UC01.json
