+++ graphwalker-model-checker/src/test/resources/json/petClinic.json
