+++ graphwalker-restful/src/test/java/org/graphwalker/restful/RestTest.java
@@ -166,7 +166,7 @@ public class RestTest extends ExecutionContext implements RestFlow {
   @Override
   public void e_Load() {
     HttpPost request = new HttpPost("http://localhost:9191/graphwalker/load");
-    FileEntity fileEntity = new FileEntity(ResourceUtils.getResourceAsFile("gw3/UC01.gw3"), ContentType.TEXT_PLAIN);
+    FileEntity fileEntity = new FileEntity(ResourceUtils.getResourceAsFile("gw3/UC01.json"), ContentType.TEXT_PLAIN);
     request.setEntity(fileEntity);
     try {
       response = HttpClientBuilder.create().build().execute(request);
