+++ graphwalker-restful/src/test/resources/gw3/UC01.json
